// client/src/app/store/collection/types/collection-actions.enum.ts
export const enum COLLECTION_ACTIONS {
  LOAD_COLLECTIONS = '[Collection] Load Collections',
  LOAD_COLLECTIONS_SUCCESS = '[Collection] Load Collections Success',
  LOAD_COLLECTIONS_FAILURE = '[Collection] Load Collections Failure',

  GET_COLLECTION = '[Collection] Get Collection',
  GET_COLLECTION_SUCCESS = '[Collection] Get Collection Success',
  GET_COLLECTION_FAILURE = '[Collection] Get Collection Failure',

  ADD_COLLECTION = '[Collection] Add Collection',
  ADD_COLLECTION_SUCCESS = '[Collection] Add Collection Success',
  ADD_COLLECTION_FAILURE = '[Collection] Add Collection Failure',

  UPDATE_COLLECTION = '[Collection] Update Collection',
  UPDATE_COLLECTION_SUCCESS = '[Collection] Update Collection Success',
  UPDATE_COLLECTION_FAILURE = '[Collection] Update Collection Failure',

  REMOVE_COLLECTION = '[Collection] Remove Collection',
  REMOVE_COLLECTION_SUCCESS = '[Collection] Remove Collection Success',
  REMOVE_COLLECTION_FAILURE = '[Collection] Remove Collection Failure',
}